+++ src/app/eth-common/module/toolbar/alethioMonitoringModule.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { IModuleDef } from "plugin-api/IModuleDef";
 import { ToolbarItem } from "@alethio/ui/lib/layout/toolbar/ToolbarItem";
 import { ITranslation } from "plugin-api/ITranslation";
-import { ExternalLink } from "plugin-api/component/ExternalLink";
+import { Link } from "plugin-api/component/Link";
 import { ProductIconButton } from "@alethio/ui/lib/layout/toolbar/ProductIconButton";
 import { AlethioMonitoringIcon } from "@alethio/ui/lib/icon/AlethioMonitoringIcon";
 
@@ -16,9 +16,9 @@ export const alethioMonitoringModule: (alethioMonitoringUrl: string) => IModuleD
     dataAdapters: [],
     getContentComponent: async () => (props) => (
         <ToolbarItem title={props.translation.get("toolbar.monitoring.label")}>
-            <ExternalLink href={alethioMonitoringUrl} rel="noopener noreferrer">
+            <Link to={alethioMonitoringUrl}>
                 <ProductIconButton Icon={AlethioMonitoringIcon} />
-            </ExternalLink>
+            </Link>
         </ToolbarItem>
     ),
     getContentProps : data => data,