+++ src/app/eth-common/module/toolbar/alethioReportsModule.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import { IModuleDef } from "plugin-api/IModuleDef";
 import { ToolbarItem } from "@alethio/ui/lib/layout/toolbar/ToolbarItem";
 import { ITranslation } from "plugin-api/ITranslation";
-import { ExternalLink } from "plugin-api/component/ExternalLink";
+import { Link } from "plugin-api/component/Link";
 import { ToolbarIconButton } from "@alethio/ui/lib/layout/toolbar/ToolbarIconButton";
 import { AlethioReportsIcon } from "@alethio/ui/lib/icon/AlethioReportsIcon";
 
@@ -16,9 +16,9 @@ export const alethioReportsModule: (alethioReportsUrl: string) => IModuleDef<IAl
     dataAdapters: [],
     getContentComponent: async () => (props) => (
         <ToolbarItem title={props.translation.get("toolbar.reports.label")}>
-            <ExternalLink href={alethioReportsUrl} rel="noopener noreferrer">
+            <Link to={alethioReportsUrl}>
                 <ToolbarIconButton Icon={AlethioReportsIcon} />
-            </ExternalLink>
+            </Link>
         </ToolbarItem>
     ),
     getContentProps : data => data,